+++ packages/icons/docs/stories/helpers/iconsArray.js
@@ -29,6 +29,15 @@ import {
     Minus,
     Avatar,
     FocusFilled,
+    Visa,
+    Amex,
+    Mastercard,
+    Discover,
+    JCB,
+    UnionPay,
+    NewCard,
+    UnknownCard,
+    DinersClub,
 } from '../../../src';
 
 export default [
@@ -181,5 +190,50 @@ export default [
         name: 'avatar',
         category: 'custom',
         icon: Avatar,
+    },
+    {
+        name: 'visa',
+        category: 'custom',
+        icon: Visa,
+    },
+    {
+        name: 'amex',
+        category: 'custom',
+        icon: Amex,
+    },
+    {
+        name: 'mastercard',
+        category: 'custom',
+        icon: Mastercard,
+    },
+    {
+        name: 'discover',
+        category: 'custom',
+        icon: Discover,
+    },
+    {
+        name: 'jcb',
+        category: 'custom',
+        icon: JCB,
+    },
+    {
+        name: 'unionPay',
+        category: 'custom',
+        icon: UnionPay,
+    },
+    {
+        name: 'dinersClub',
+        category: 'custom',
+        icon: DinersClub,
+    },
+    {
+        name: 'newCard',
+        category: 'custom',
+        icon: NewCard,
+    },
+    {
+        name: 'unknownCard',
+        category: 'custom',
+        icon: UnknownCard,
     },
 ];